Prepare CtsUsbTests for instant apps
The only thing the test does is to compare the serial number. This is
not possible in instant app mode as the instant app cannot get the
permission.

Hence add a test that ensures that the call fails.

Further fix the host side test to correctly collect errors from the
device side sub-test.

Fixes: 79376967
Test: cts-tradefed run cts-dev CtsUsbTests
      cts-tradefed run cts-instant CtsUsbTests

Change-Id: I62ad8c71d9abcafffbf254f1f93cacc5db4f3dae