Allow androidx path for explicit runners in cts
Runners specified explicitly are allowed to be in the new
package for Android runner.

And

Only check explicitly define runners in the presubmit check

Avoid not being happy when no runner is defined. It's
perfectly valid for the tests to let the runner dynamically
find the proper runner.

Test: make cts && cts-tradefed.sh run cts-unit-tests
Bug: 123431359
Merged-In: Ia6696433e6cc9eafa7bab6959befb76d30aed0da
(merged from master)

Change-Id: I5b3675450dfcdfac6219f5d94867f2ef833a5ef5