Add per-user methods to BackupUtils
To support adding CTS/GTS tests for multi-user backup/restore, create
versions of existing helpers that act on a specified user.

New versions of helpers are created instead of changing the existing
ones as user-specific methods are not available on older platforms and
we have tests on older platforms currently using these helpers.

Bug: 122305663
Test: 1) atest CtsBackupHostTestCases
2) atest GtsBackupHostTestCases
3) atest ManagedProfileKeyValueBackupRestoreHostSideTest

Change-Id: I18b5710d971feabafec9e382d3e2252aabac8ae0