Fix logcat parsing flakiness
When checking that a set of logs appear in order, we previously wrongly
required that all logs had to appear in the same "adb logcat"
invocation. This causes flakiness as logs depend on operations that
occur at different times, so the first logs we look for might have been
gone by the time the later logs are written. Instead, allow the logs we
look for to span different "adb logcat" invocations, as we only care about
order.

Bug: 129476019
Test: 1) atest CtsBackupTestCases
2) atest CtsBackupHostTestCases
3) atest GtsBackupTestCases
4) atest GtsBackupHostTestCases

Change-Id: I304dc74df63959afd118021358c322685b9aa7ca