[Cherry pick] Custom ResultReporter for STS
Currently, the retry of a pi sts build will have a suite_plan name of "sts-retry" instead of the original `suite_plan` of the test result being retried.

This is a regression from N, O behavior.

This has the potential to:
* Break APFE validation when we start approving based on variant/representative and suite-plan combination
* Break metrics

Bug: 126423649
Test: mm sts-harness-tradefed-tests && ./run_tests.sh
Test: Ran retry using tradefed and manually validated test_result.xml
Change-Id: Ieb2053e47b19097984ec16ec178add5a04f98e6f
Merged-In: Ieb2053e47b19097984ec16ec178add5a04f98e6f