Use location_mode instead of location_providers_allowed
The LocationCheck failed for devices without GPS and NETWORK location provider
when GSI was installed. Instead of checking if location_providers_allowed
has either gps and/or network, this CL changed to check if location_mode
is on or not.

Bug: 140610515
Test: run cts-on-gsi -m CtsLocationTestCases
Change-Id: I18a878e6c6d89e3fac0084e50d8db8f5d730107d
(cherry picked from commit fb4f626ab52318c737592355a717328f8f03c84e)