allow inserting arbitary attributes into Result XML tag
This change enables arbitrary key-value pairs to be added to
the "Result" XML tag in CTS result XML as attributes and their
corresponding values.

Also included a minor fix to the Java version detection in
cts-tradefed: the previous implementation assumed that the
version number will be printed on the first line.

Bug: 148189589
Test: cts-tradefed run cts -m CtsGestureTestCases \
        --result-attribute display_mode 0
Change-Id: I5a2facbb415f879cc2f57792e5ddea35dede58b9