allow inserting arbitary attributes into Result XML tag
This change enables arbitrary key-value pairs to be added to
the "Result" XML tag in CTS result XML as attributes and their
corresponding values.

Also included a minor fix to the Java version detection in
cts-tradefed: the previous implementation assumed that the
version number will be printed on the first line.

Bug: 148189589
Change-Id: Ia773a1353d0f93e8357649458421a519fd2d52cb
Test: none (cherry-pick)